move constant editor
